fix(letters): restore inline subscribe UI
- replace the modal subscribe flow with the lighter inline design
- disable the submit button for empty or invalid email values
- switch the subscribe icons to lucide-react and keep toast-backed feedback
